ISPN-1321 Moving test and excluding generated index from git
