+++ .gitignore
@@ -32,4 +32,6 @@ rhq-plugin.xml
 core/src/main/resources/schema
 # Compiled python files
 *.pyc
+# Jandex generated indexes
+*.idx
